fix(desktop): keep settings URL validation browser-safe
